+++ core/security/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
     alias(libs.plugins.bitnagil.android.library)
-    alias(libs.plugins.bitnagil.android.hilt)
 }
 
 android {
